Updated the bit width of Ip/Op data to 32 bits. Reorganized code in app and drivers. Real time convolution with overlap add working at O3.

+++ src/convolution.c
@@ -1,7 +1,7 @@
 #include "convolution.h"
 
 
-void convolve_16x16(uint16_t *x,uint16_t *h, uint16_t *y)
+void convolve_16x16(int32_t *x,int32_t *h, int32_t *y)
 {
     uint8_t i,j;
